backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
chromadb
pymupdf
python-dotenv
openai
sentence-transformers
google-generativeai
requests 